+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.7.0
+
+### Changed
+* Changed implementation of Entity to create, edit and delete it at Celery.
+* Changed to show unauthorized entity on the dashboard
+
 ## v2.6.0
 
 ### Added